+++ src/module-registrar.cc
@@ -32,6 +32,7 @@
 #include "agent.hh"
 #include "domain-registrations.hh"
 #include "eventlogs/events/eventlogs.hh"
+#include "exceptions/bad-configuration.hh"
 #include "module-toolbox.hh"
 #include "nat/nat-traversal-strategy.hh"
 #include "registrar/binding-parameters.hh"
@@ -396,23 +397,29 @@ void ModuleRegistrar::declareConfig(GenericStruct& moduleConfig) {
 	    {
 	        DurationS,
 	        "max-expires",
-	        "Maximum expire time for a REGISTER.",
+	        "Maximum expiry value for a REGISTER.",
 	        "86400",
 	    },
 	    {
 	        DurationS,
 	        "min-expires",
-	        "Minimum expire time for a REGISTER.",
+	        "Minimum expiry value for a REGISTER.",
 	        "60",
 	    },
 	    {
-	        Integer,
+	        DurationMS,
+	        "default-expires",
+	        "Default expiry value to be used if no value has been found in the request headers or in 'Contact' header "
+	        "parameters.",
+	        "10min",
+	    },
+	    {
+	        DurationS,
 	        "force-expires",
-	        "Set a value that will override expire times given by the "
-	        "REGISTER requests. A null or negative value disables "
-	        "that feature. If it is enabled, max-expires and min-expires "
-	        "will not have any effect.",
-	        "-1",
+	        "Set a value that will override expiry values indicated in a 'REGISTER' request. A null or negative value "
+	        "disables this feature. If enabled, 'max-expires', 'min-expires' and 'default-expires' will not have any "
+	        "effect.",
+	        "0",
 	    },
 	    {
 	        String,
@@ -439,7 +446,6 @@ void ModuleRegistrar::declareConfig(GenericStruct& moduleConfig) {
 	        "The redis backend is recommended, the internal being more adapted to very small deployments.",
 	        "internal",
 	    },
-
 	    // Redis config support
 	    {
 	        String,
@@ -566,18 +572,21 @@ void ModuleRegistrar::onLoad(const GenericStruct* mc) {
 	// replace space-separated to comma-separated since sofia-sip is expecting this way
 	std::replace(mServiceRoute.begin(), mServiceRoute.end(), ' ', ',');
 
-	int forcedExpires = mc->get<ConfigInt>("force-expires")->read();
-	if (forcedExpires <= 0) {
-		mMaxExpires =
-		    chrono::duration_cast<chrono::seconds>(mc->get<ConfigDuration<chrono::seconds>>("max-expires")->read())
-		        .count();
-		mMinExpires =
-		    chrono::duration_cast<chrono::seconds>(mc->get<ConfigDuration<chrono::seconds>>("min-expires")->read())
-		        .count();
-		if (mMaxExpires < mMinExpires) LOGF("Registrar 'max-expires' must be equal to or greater than 'min-expires'");
+	const auto forcedExpires =
+	    chrono::duration_cast<chrono::seconds>(mc->get<ConfigDuration<chrono::seconds>>("force-expires")->read());
+	if (forcedExpires <= 0s) {
+		const auto maxExpires = mc->get<ConfigDuration<chrono::seconds>>("max-expires");
+		mMaxExpires = chrono::duration_cast<chrono::seconds>(maxExpires->read()).count();
+		const auto minExpires = mc->get<ConfigDuration<chrono::seconds>>("min-expires");
+		mMinExpires = chrono::duration_cast<chrono::seconds>(minExpires->read()).count();
+		if (mMaxExpires < mMinExpires)
+			throw BadConfiguration{maxExpires->getCompleteName() + " must be equal or greater than " +
+			                       minExpires->getCompleteName()};
+		mDefaultExpires = mc->get<ConfigDuration<chrono::milliseconds>>("default-expires")->read();
 	} else {
-		mMaxExpires = forcedExpires;
-		mMinExpires = forcedExpires;
+		mMaxExpires = forcedExpires.count();
+		mMinExpires = forcedExpires.count();
+		mDefaultExpires = forcedExpires;
 	}
 
 	mStaticRecordsFile = mc->get<ConfigString>("static-records-file")->read();
@@ -837,18 +846,6 @@ void ModuleRegistrar::onRequest(shared_ptr<RequestSipEvent>& ev) {
 		return;
 	}
 	const auto* expires = sip->sip_expires;
-	const auto mainDelta = normalizeMainDelta(expires, mMinExpires, mMaxExpires);
-	if (!checkHaveExpire(sip->sip_contact, mainDelta)) {
-		SLOGD << "No global or local expire found in at least one contact";
-		reply(ev, 400, "Invalid request");
-		return;
-	}
-	for (auto contact = sip->sip_contact; contact != nullptr; contact = contact->m_next) {
-		if (!isValidSipUri(contact->m_url)) {
-			reply(ev, 400, "Invalid contact");
-			return;
-		}
-	}
 	if (!usageOfWildcardContactIsCorrect(sip->sip_contact, expires)) {
 		reply(ev, 400, "Invalid usage of wildcard '*' in Contact header field");
 		return;
@@ -856,6 +853,13 @@ void ModuleRegistrar::onRequest(shared_ptr<RequestSipEvent>& ev) {
 	if (numberOfContactHeaders(sip->sip_contact) > mMaxContactsPerRegistration) {
 		reply(ev, 403, "Too many contacts in REGISTER");
 		return;
+	}
+
+	auto mainDelta = normalizeMainDelta(expires, mMinExpires, mMaxExpires);
+	if (!checkHaveExpire(sip->sip_contact, mainDelta)) {
+		const auto defaultExpires = chrono::duration_cast<chrono::seconds>(mDefaultExpires).count();
+		SLOGD << "No expiry value found in request, using default expiry value: " << defaultExpires << "s";
+		mainDelta = defaultExpires;
 	}
 
 	// Use path as a contact route in all cases